Revert "feat hard work..."
This reverts commit 0c4a54fef9cae4c2a37caf9a20fb819d6430b92c.

+++ Domain/Entities/DesignCategory.cs
@@ -7,7 +7,7 @@ namespace Domain.Entities
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long DesignCategoryId { get; set; }
+        public int DesignCategoryId { get; set; }
         public string DesignCategoryName { get; set; }
         public List<Design> Designs { get; set; }
     }
